% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publr_color.R
\name{publr_color}
\alias{publr_color}
\title{This function returns color palettes}
\usage{
publr_color(color_palette = "cpxq04", display = FALSE)
}
\arguments{
\item{color_palette}{Name of color palette to return}

\item{display}{If TRUE, a plot of n palettes is shown for comparison}
}
\description{
This function returns color palettes
}
