% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publr_theme.R
\name{publr_theme}
\alias{publr_theme}
\title{publR ggplot2 theme: A easy customizable theme for publishing ggplot2 figures}
\usage{
publr_theme(font_family = "sans", axis_line_width = 0.2,
  font_size_labels = 12, font_size_axis = 10, text_color = "black")
}
\arguments{
\item{font_family}{The font family used in all text of the figure}

\item{axis_line_width}{Size of axis lines and border around legend}

\item{font_size_labels}{Size of fonts in axis labels}

\item{font_size_axis}{Size of fonts in axis text}

\item{text_color}{Color of text in figure. The same color will be applied to all text}
}
\description{
publR ggplot2 theme: A easy customizable theme for publishing ggplot2 figures
}
