% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{# Input a non-truncated p-value, the function returns a rounded p-value to p<1/10^digits or}
\usage{
pval(pval, digits = 3, flag = FALSE, plot = FALSE)
}
\arguments{
\item{pval}{A p-value to be truncated}

\item{digits}{To how many decimal points the p-value should be truncated, e.g. 3 digits returns p < 0.001 as lowest threshold}

\item{flag}{If TRUE, a flag is returned instead of p-values. p-values 0.01-0.05, "*"; 0.001-0.01, "**"; 0.0001-0.001, "***"; < 0.0001, "****".}

\item{plot}{If TRUE, p-value is written to be inserted in e.g. ggplot2 figures in the form bquote(italic(p)<.(lowp)), elese returned as character for printing in text.}
}
\value{
Truncetad p-values or flag to be used in reporting e.g. rmarkdown documents or figures.
}
\description{
# Input a non-truncated p-value, the function returns a rounded p-value to p<1/10^digits or
}
